requirements.txt
phidata
openai
pandas
plotly
streamlit
duckduckgo-search
python-dotenv
pydantic
sqlalchemy
psycopg2-binary
stripe
requests
